ignore rm of cidfile if it not exist
apparently some versions of podman delete also the cidfile during podman
rm and some not. That leads to situation that in some OSes (e.g., c9s)
the cleaning function fails on podman rm, as the cid_file is removed
already removed at that point. This commit fixes this behaviour.